requirements.txt
pandas>=1.0.5
pytest>=6.0.1
pytz>=2020.1
requests>=2.24.0
plumbum>=1.6.9
pyyaml>=5.3.1